tests/wpt/web-platform-tests/app-history/navigate/navigate-info-and-state.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async t => {
  let navInfo = { nav : "info" };
  let navState = { statevar: "state" };
  let onnavigated_called = false;
  appHistory.onnavigate = t.step_func(e => {
    onnavigated_called = true;
    assert_equals(e.info, navInfo)
  });
  await appHistory.navigate("#1", { info: navInfo, state: navState });
  assert_true(onnavigated_called);
  assert_not_equals(appHistory.current.getState(), navState);
  assert_equals(appHistory.current.getState().statevar, navState.statevar);
}, "navigate() with info and state");
</script>
